feature/새 상품 등록 페이지 컴포넌트화, fix/마이 페이지에 상품 등록 페이지로 이동되지 않는 문제 수정
