+++ src/components/EditProduct.jsx
@@ -1,0 +1,3 @@
+export default function EditProduct() {
+  return;
+}
