+++ src/components/SoldItem.jsx
@@ -1,4 +1,5 @@
 import PropTypes from "prop-types";
+import { Link } from "react-router-dom";
 
 SoldItem.propTypes = {
   item: PropTypes.shape().isRequired,
@@ -9,12 +10,19 @@ export default function SoldItem({ item }) {
   const productSelling = item.quantity - item.buyQuantity > 0 ? true : false;
   return (
     <section className="flex gap-5 py-3 items-center">
-      <img
-        src={"https://11.fesp.shop" + item.mainImages[0].path}
-        className=" w-[100px] h-[100px] aspect-square object-cover rounded-md"
-      />
+      <Link
+        to={`/product/${item._id}`}
+        className=" w-[100px] h-[100px] aspect-square object-cover"
+      >
+        <img
+          src={"https://11.fesp.shop" + item.mainImages[0].path}
+          className="aspect-square object-cover rounded-md"
+        />
+      </Link>
       <div className="text-sm">
-        <p className="font-semibold break-keep">{item.name}</p>
+        <Link to={`/product/${item._id}`}>
+          <p className="font-semibold break-keep">{item.name}</p>
+        </Link>
         <p className="text-xs  text-gray5 py-1 pb-3">
           {productSelling
             ? "현재 판매 중"
