+++ src/routes.jsx
@@ -10,11 +10,11 @@ const ProductDetailPage = lazy(() => import("@pages/market/ProductDetailPage"));
 const ProductNewPage = lazy(() => import("@pages/market/ProductNewPage"));
 const ReviewPage = lazy(() => import("@pages/market/ReviewPage"));
 const PhotoReviewPage = lazy(() => import("@pages/market/PhotoReviewPage"));
-const ProductMyReviewPage = lazy(
-  () => import("@pages/market/ProductMyReviewPage")
+const ProductMyReviewPage = lazy(() =>
+  import("@pages/market/ProductMyReviewPage")
 );
-const ProductNewReviewPage = lazy(
-  () => import("@pages/market/ProductNewReviewPage")
+const ProductNewReviewPage = lazy(() =>
+  import("@pages/market/ProductNewReviewPage")
 );
 const BookmarkPage = lazy(() => import("@pages/user/BookmarkPage"));
 const MyPage = lazy(() => import("@pages/user/MyPage"));
@@ -35,8 +35,8 @@ const EditProfilePage = lazy(() => import("@pages/user/EditProfilePage"));
 const MyPostPage = lazy(() => import("@pages/user/MyPostPage"));
 const SearchBestPage = lazy(() => import("@pages/market/SearchBestPage"));
 const SearchNewPage = lazy(() => import("@pages/market/SearchNewPage"));
-const SearchSeasonalPage = lazy(
-  () => import("@pages/market/SearchSeasonalPage")
+const SearchSeasonalPage = lazy(() =>
+  import("@pages/market/SearchSeasonalPage")
 );
 const Layout = lazy(() => import("@components/layout"));
 
@@ -87,8 +87,14 @@ const router = createBrowserRouter(
             { path: "signup", element: <SignupPage /> },
             { path: "login", element: <LoginPage /> },
             { path: "mypage", element: <MyPage /> },
-            { path: "profile", element: <ProfilePage /> },
-            { path: "profile/edit", element: <EditProfilePage /> },
+            {
+              path: "profile",
+              children: [
+                { index: true, element: <ProfilePage /> },
+                { path: "edit", element: <EditProfilePage /> },
+              ],
+            },
+
             { path: "bookmarks", element: <BookmarkPage /> },
             { path: "recent", element: <RecentPage /> },
             { path: "sale", element: <SalePage /> },
